* 'master' of https://bitbucket.org/lei-isep/switch-2017-g003:
  Cleaning code smells in controllers
  Corrected the tests that were using the variable int EffortCost of the ProjectCollaborator Class.
  Changed the effort unit in XML file from PERSON_MONTH to PM, in accordance to the new constant in the EffortUnit enum Class.
  Changed the type of the variable costPerEffort from int to Double in the ProjectCollaborator Class
  Refactor: moved the classes FileUtils, LoadProjectData and LoadUserData to a new package called loadFile
  Corrected the assertEquals that were using the getProjectBudget method.
  Changed the type of the budget variable of the Project Class to Double.
  Refactored the constant PERSON_MONTH to PM in the EffortUnit Enum Class.
  Debugged class variable initialization in create project UI. Variables remained unchanged after the first project creation, including the boolean to enter the creation loop
  Updated US203 Diagram (Changed UnfinishedTasks to ToDoTasks..)
  added @Transactional annotation to prevent lazy initialization crashes
  Completed XML parser for Project Data
  Created US203 Sequence Diagram
  Fixed load user data to prevent null fields in User ID and Function
  US205MarkTaskAsFinishedCollaboratorTest: add test of setters
  Added missing set for the controllers.
  US367MarkFinishedTaskAsUnfinishedControllerTest: add test for setter
  Corrected project file to parse.
  US204v2createRequestAddCollaboratorToTaskTeamControllerTest: add tests for setters

+++ src/main/java/project/model/Project.java
@@ -31,7 +31,7 @@ public class Project implements Serializable{
 	private String description;
 	@Enumerated(EnumType.STRING)
 	private EffortUnit effortUnit;
-	private int budget;
+	private double budget;
 	private Calendar startdate;
 	private Calendar finishdate;
 
@@ -105,7 +105,7 @@ public class Project implements Serializable{
 		this.description = description;
 	}
 
-	public int getBudget() {
+	public double getBudget() {
 		return budget;
 	}
 
@@ -345,14 +345,14 @@ public class Project implements Serializable{
 	 * 
 	 * @return Budget (int) of Project
 	 */
-	public int getProjectBudget() {
+	public double getProjectBudget() {
 		return this.budget;
 	}
 
 	/**
 	 * Set Project budget
 	 */
-	public void setProjectBudget(int newBudget) {
+	public void setProjectBudget(double newBudget) {
 		this.budget = newBudget;
 	}
